style: move persisten-model to ~/models
